platzi/requirements.txt
# Instalacion de Dependencias para el proyecto

# Framework principal
Django

# Para consumir APIs externas
requests
# Django REST Framework para crear APIs
djangorestframework

# Para documentación automática de la API (opcional pero recomendado)
drf-spectacular

# Para validación adicional y utilidades
python-decouple  # Para variables de entorno
Pillow  # Si necesitas manejo de imágenes
psycopg2-binary  # Para conexión a PostgreSQL

# Para instalar las dependencias, ejecutar:
# pip install -r requirements.txt